fix(daemon): deduplicate StartRecording into a shared flow with per-platform ffmpeg capture commands and unified childreap.Reap exit handling
